Correcting and polishing
Corrected segmentation fault in the case where we need to sort only 2 numbers, ft_error now displays the error message on stderr

+++ libft/SRC/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 16:42:47 by mely-pan          #+#    #+#             */
-/*   Updated: 2024/05/13 19:09:42 by mely-pan         ###   ########.fr       */
+/*   Updated: 2024/12/18 20:04:38 by mely-pan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
